requirements.txt
fastapi
uvicorn
pydantic
email-validator
python-multipart
python-jose[cryptography]
passlib[bcrypt]